+++ test/regress/test-s1.bash
@@ -16,6 +16,13 @@ if [ ! -d $RESULTS_DIR ]; then
 fi
 
 ## Tests
+
+# Initiate backups (full, diff, incr)
+echo "Initiate backups (full, diff, incr)"
+sudo -iu postgres pgbackrest --stanza=my_stanza backup --type=full --repo1-retention-full=1
+sudo -iu postgres pgbackrest --stanza=my_stanza backup --type=diff
+sudo -iu postgres pgbackrest --stanza=my_stanza backup --type=incr
+
 # --list
 echo "--list"
 $PLUGIN_PATH/check_pgbackrest --list > $RESULTS_DIR/list.out
@@ -24,23 +31,22 @@ $PLUGIN_PATH/check_pgbackrest --list > $RESULTS_DIR/list.out
 echo "--version"
 $PLUGIN_PATH/check_pgbackrest --version > $RESULTS_DIR/version.out
 
-# --service=retention missing arg
-echo "--service=retention missing arg"
-$PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=retention > $RESULTS_DIR/retention-missing-arg.out 2>&1
-
 # --service=retention --retention-full
 echo "--service=retention --retention-full"
-sudo -iu postgres pgbackrest --stanza=my_stanza backup --type=full --repo1-retention-full=1
 $PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=retention --retention-full=1 | cut -f1 -d"|" > $RESULTS_DIR/retention-full.out
 
 # --service=retention --retention-age
 echo "--service=retention --retention-age"
 $PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=retention --retention-age=1h | cut -f1 -d"|" > $RESULTS_DIR/retention-age.out
 
+# --service=retention --retention-age-to-full
+echo "--service=retention --retention-age-to-full"
+$PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=retention --retention-age-to-full=1h | cut -f1 -d"|" > $RESULTS_DIR/retention-age-to-full.out
+
 # --service=retention fail
 echo "--service=retention fail"
-sudo -iu postgres psql -c "SELECT pg_sleep(1);" > /dev/null 2>&1
-$PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=retention --retention-full=2 --retention-age=1s | cut -f1 -d"|" > $RESULTS_DIR/retention-fail.out
+sudo -iu postgres psql -c "SELECT pg_sleep(2);" > /dev/null 2>&1
+$PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=retention --retention-full=2 --retention-age=1s --retention-age-to-full=1s | cut -f1 -d"|" > $RESULTS_DIR/retention-fail.out
 
 # --service=archives missing arg
 echo "--service=archives missing arg"
@@ -58,6 +64,7 @@ $PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=archives --repo-path=
 
 # --service=archives --latest-archive-age-alert
 echo "--service=archives --latest-archive-age-alert"
+sudo -iu postgres psql -c "SELECT pg_sleep(2);" > /dev/null 2>&1
 $PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=archives --repo-path=/var/lib/pgbackrest/archive --latest-archive-age-alert=1h | cut -f1 -d"-" > $RESULTS_DIR/archives-age-alert-ok.out
 $PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=archives --repo-path=/var/lib/pgbackrest/archive --latest-archive-age-alert=1s | cut -f1 -d"-" > $RESULTS_DIR/archives-age-alert-ko.out